requirements.txt
Django
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-js-asset==0.1.1
django-pure-pagination==0.3.0
django-simple-captcha==0.5.5
djangorestframework==3.6.3
future==0.16.0
httplib2==0.10.3
Markdown==2.6.8
numpy==1.12.1
olefile==0.44
Pillow==4.2.0
PyMySQL==0.7.11
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
django-ckeditor
requests
django-rest-swagger